修复修改可选频率后“从 int64_t 收缩到 int”的问题

+++ include/utils/scope_key.h
@@ -20,7 +20,7 @@ struct ScopeKey {
     int window{0};
 
     ScopeKey() = default;
-    ScopeKey(std::string c, int f, int w)
+    ScopeKey(std::string c, int64_t f, int w)
         : code(std::move(c)), frequency(f), window(w) {}
 
     /**
